requirements.txt
nltk
PyMuPDF
python-docx
transformers
